Update windows-defender-advanced-threat-protection.md
"Configure and manage..." currently the link title is shown as "Cconfigure and manage...." I have updated the link title for the onboard.md in this fork.
Thank you for all of your hard work.  Please did not see this "nit-picking" but as someone is actually reading and using your efforts - just paying attention!  Great job Y'all!!
-JED